requirements.txt
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2
mlflow==2.10.2
dvc==3.38.1
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
evidently==0.4.22
pytest==8.0.1
pytest-cov==4.1.0
httpx==0.27.0
python-dotenv==1.0.1
PyYAML==6.0.1
huggingface_hub==0.21.0